add: typeguard for middleware ts types/fix-all#239

+++ src/server/middleware.ts
@@ -1,5 +1,6 @@
 import { REVALIDATION_THRESHOLD } from '@/config/env'
 import { isObject } from '@/utils/typeguards'
+import { Report } from '@prisma/client'
 import { TRPCError } from '@trpc/server'
 import { middleware } from './trpc'
 
@@ -19,25 +20,27 @@ export const isAuthedMiddleWare = middleware(({ next, ctx }) => {
 })
 
 export const shouldRevalidateMiddleWare = middleware(async ({ next, ctx, rawInput }) => {
-  const count = await ctx.prisma.report.count({
-    where: {
-      request_type: 'REVALIDATE',
-      body: {
-        equals: rawInput.body,
-      },
-    },
-  })
-
-  if (count + 1 >= REVALIDATION_THRESHOLD) {
-    await ctx.prisma.report.deleteMany({
+  if (isObject<Pick<Report, 'body'>>(rawInput)) {
+    const count = await ctx.prisma.report.count({
       where: {
+        request_type: 'REVALIDATE',
         body: {
           equals: rawInput.body,
         },
       },
     })
 
-    await ctx.res.revalidate(`/details/${rawInput.body}`)
+    if (count + 1 >= REVALIDATION_THRESHOLD) {
+      await ctx.prisma.report.deleteMany({
+        where: {
+          body: {
+            equals: rawInput.body,
+          },
+        },
+      })
+
+      await ctx.res.revalidate(`/details/${rawInput.body}`)
+    }
   }
 
   return next()
